refactor(files): split FilePanel into focused modules
The 949-line FilePanel.tsx is split into a FilePanel/ folder:
- index.tsx     — orchestrator: explorer state, file ops, menu, mode switch
- TreeBrowser.tsx — explorer UI (toolbar, tree, rows, inline name input)
- FileViewer.tsx  — load/error/loading wrapper around the editor
- FileEditor.tsx  — the textarea + highlight editor
- ViewerHeader.tsx — header shared by viewer + editor
- tree.ts         — tree model types + buildTree/filterTree/duplicatePath/errMsg
- helpers.test.ts — moved here, now imports ./tree

Pure refactor, no behavior change. Explorer state stays in the orchestrator
(not TreeBrowser) so it survives a round-trip through the editor — notably
pendingDirs, which tracks freshly-created empty folders git won't list.
`./FilePanel` still resolves (folder index), so RightPanel's import is
unchanged. tsc + build + all 122 tests pass.
