+++ src/components/RightPanel/FilePanel/helpers.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "vitest";
-import { buildTree, duplicatePath } from "./FilePanel";
-import { joinPath, parentDir } from "../../util/format";
-import type { WorktreeFile } from "../../api";
+import { buildTree, duplicatePath } from "./tree";
+import { joinPath, parentDir } from "../../../util/format";
+import type { WorktreeFile } from "../../../api";
 
 const f = (path: string): WorktreeFile => ({ path, status: null, additions: 0, deletions: 0 });
